+++ package.json
@@ -1,50 +1,47 @@
 {
-  "name": "monday-integration-quickstart-app",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "start": "npm run stop && concurrently \"npm run server\" \"npm run expose\"",
-    "build": "react-scripts build",
-    "expose": "ngrok http 8301",
-    "server": "react-scripts start",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "stop": "kill-port 8301 && kill-port 4049 && kill-port 4040"
-  },
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "monday-sdk-js": "0.0.16",
-    "pdf-parse": "^1.1.1",
-    "pdfjs-dist": "^2.13.216",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
-    "react-scripts": "3.4.0"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@types/monday-sdk-js": "^0.1.5",
-    "@types/pdf-parse": "^1.1.1",
-    "@types/react": "^18.0.8",
-    "@types/react-dom": "^18.0.0",
-    "concurrently": "^5.2.0",
-    "cross-port-killer": "^1.2.1",
-    "ngrok": "^3.2.7",
-    "typescript": "^4.6.3"
-  }
+	"name": "monday-integration-quickstart-app",
+	"version": "0.0.1",
+	"private": true,
+	"scripts": {
+		"start": "npm run stop && concurrently \"npm run server\" \"npm run expose\"",
+		"build": "react-scripts build",
+		"expose": "ngrok http 8301",
+		"server": "react-scripts start",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject",
+		"stop": "kill-port 8301 && kill-port 4049 && kill-port 4040"
+	},
+	"dependencies": {
+		"@testing-library/jest-dom": "^4.2.4",
+		"@testing-library/react": "^9.5.0",
+		"@testing-library/user-event": "^7.2.1",
+		"monday-sdk-js": "0.0.16",
+		"react": "^16.13.0",
+		"react-dom": "^16.13.0",
+		"react-scripts": "3.4.0"
+	},
+	"eslintConfig": {
+		"extends": "react-app"
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"devDependencies": {
+		"@types/monday-sdk-js": "^0.1.5",
+		"@types/react": "^18.0.8",
+		"@types/react-dom": "^18.0.0",
+		"concurrently": "^5.2.0",
+		"cross-port-killer": "^1.2.1",
+		"ngrok": "^3.2.7",
+		"typescript": "^4.6.3"
+	}
 }
